requirements.txt
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
Pillow==9.0.1
scikit_learn==1.0.2
seaborn==0.11.2
tqdm==4.62.3
